Fixed the issue with the order of piles.
The issue was that pile[0] will have 4 cards and each card was replacing the value at pile[0] with the new value. To fix this i added another Stack to put the old values in this new stack and popped the value from the pile stack. than added them again to maintain the order.